% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_plot.R
\name{caret_plot}
\alias{caret_plot}
\title{Plot Caret Grid Search Hyper-parameter Tuning Results}
\usage{
caret_plot(
  x,
  sqrt = FALSE,
  marg1 = FALSE,
  marg2 = FALSE,
  col = NULL,
  row = NULL,
  facet = NULL
)
}
\arguments{
\item{x}{An object of class train.}

\item{sqrt}{Boolean value indicating whether assessment metrics should be
adjusted via a square root transformation. Default is FALSE.}

\item{marg1}{Boolean value indicating whether to aggregate performance
down to 1 dimension of hyper-parameter and provide the corresponding
data.frames and line-plots for assessment. Default is FALSE.}

\item{marg2}{Boolean value indicating whether to aggregate performance
down to 2 dimensions of hyper-parameter and provide the corresponding
data.frames and heatmaps for assessment. Default is FALSE.}

\item{col}{Name of the variable to plot on the columns of heatmaps. Only
relevant for heatmaps of 3 or more dimensions. Default is NULL.}

\item{row}{Name of the variable to plot on the rows of heatmaps. Only
relevant for heatmaps of 3 or more dimensions. Default is NULL.}

\item{facet}{Name of the variable to plot as the facets of heatmaps. Only
relevant for heatmaps of 4 dimensions. Default is NULL.}
}
\value{
A list of caret training performance data.frames, heatmaps, and
  plots.
}
\description{
This function uses caret grid training results to generate
  performance data.frames, heatmaps, and other plots for comparing the
  performance of models across their hyper-parameters and evaluating
  interactions between different model hyper-parameters.
}
\examples{
set.seed(123)
fit_control <- caret::trainControl(method = "cv", number = 10)
gbm_grid <- expand.grid(
  interaction.depth = c(1, 4), n.trees = c(15, 150),
  shrinkage = c(0.05, 0.1), n.minobsinnode = 10
)
x <- caret::train(factor(Species) ~ .,
  method = "gbm", tuneGrid = gbm_grid,
  trControl = fit_control, data = iris
)
p <- caret_plot(x, sqrt = FALSE, col = "n.trees", marg1 = TRUE, marg2 = TRUE)
p
}
