% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvip.R
\name{ggvip}
\alias{ggvip}
\title{Variable Importance GGPlot}
\usage{
ggvip(x, scale = F, sqrt = T, metrics = "both", ...)
}
\arguments{
\item{x}{An object of class randomForest.}

\item{...}{Arguments to be passed to the generic plot function, such as
\link{graphical parameters}. Generic plot will accept many arguments
including the following:

\code{main}: an overall title for the plot: see \link{title}.

\code{sub}: a sub title for the plot: see \link{title}.

\code{xlab}: a title for the x axis: see \link{title}.

\code{ylab}: a title for the y axis: see \link{title}.}
}
\value{
A plot of a 'carsimr' object with cells of white open spaces, blue
  cars, and red cars. Plotted lines separate and distinguish unique rows and
  columns.
}
\description{
Dotchart of variable importance as measured by a Random Forest
}
\examples{
rf <- randomForest(factor(Species) ~ ., importance = T, data = iris)
ggvip(rf, scale = F, sqrt = T)
}
