% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtry_pdp_compare.R
\name{mtry_pdp_compare}
\alias{mtry_pdp_compare}
\title{Mtry Tune via PDPs}
\usage{
mtry_pdp_compare(
  formula,
  data = NULL,
  mvec,
  var_vec,
  trim = 0.1,
  trellis = TRUE,
  which_class = 2L,
  prob = TRUE,
  ...
)
}
\arguments{
\item{formula}{an object of class "\link{formula}" (or one that can be
coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame containing the variables in the model.
By default the variables are taken from the environment which randomForest
is called from.}

\item{mvec}{Optional vector argument for defining choices of mtry to have the
function consider. Should be a vector of integers between 1 and the total
number of predictor variables in the model. Or it can be a vector of
proportions (strictly less than 1) of the number of predictor variables.}

\item{var_vec}{Optional vector argument for reducing the number of
variables to consider and compare. Elements should be characters that
match column names from the data used to generate the model x.}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from
each end of an individual PDP dataset before the trim-range is computed.
The default of 0.1 will be used when values of trim outside that range
are given.}

\item{trellis}{Logical indicating whether or not to generate trellis plots
as output for comparing PDPs. Default is TRUE.}

\item{which_class}{Integer specifying which column of the matrix of
predicted probabilities to use as the "focus" class. Default is to
use the first class. Only used for classification problems.}

\item{prob}{Logical indicating whether or not partial dependence for
classification problems should be returned on the probability scale,
rather than the centered logit. If FALSE, the partial dependence function
is on a scale similar to the logit. Default is TRUE.}

\item{...}{Other parameters to pass to the randomForest function.}
}
\value{
A list of data.frames, useful plots, and forest objects for user
  evaluations of the randomForest hyperparameter mtry. This includes a
  list of partial dependence plots with adjusted y-axes so all PDPs
  are on an identical scale. This also contains comparative facet plots
  and PDP importance values for assessing true effect of predictors
  on response.
}
\description{
This function builds randomForest algorithms, generates PDPs
  and combines them across different models. Outputs a list of data.frames
  and useful plots for user evaluations of the randomForest
  hyperparameter mtry. This also contains PDP-derived importance values for
  assessing effect of predictors on response.
}
\examples{
m <- mtry_pdp_compare(Petal.Length ~ ., data = iris)
m
}
