% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_plot_breaks.R
\name{nice_plot_breaks}
\alias{nice_plot_breaks}
\title{Generates Nice Plot Breaks and Limits from Data Range}
\usage{
nice_plot_breaks(vec)
}
\arguments{
\item{x}{An object of class train.}

\item{sqrt}{Boolean value indicating whether assessment metrics should be
adjusted via a square root transformation. Default is FALSE.}
}
\value{
A list of caret training performance data.frames, heatmaps, and
  plots.
}
\description{
This function uses caret grid training results to generate
  performance data.frames, heatmaps, and other plots for comparing the
  performance of models across their hyper-parameters and evaluating
  interactions between different model hyper-parameters.
}
\examples{
b <- nice_plot_breaks(x)
}
