% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_compare.R, R/pdp_compare_bivariate.R
\name{pdp_compare}
\alias{pdp_compare}
\title{Small Multiple PDPs and Importance Metrics}
\usage{
pdp_compare(
  x,
  var_vec,
  scale = FALSE,
  sqrt = TRUE,
  trim = 0.1,
  trellis = TRUE,
  which_class = 2L,
  prob = TRUE,
  ...
)

pdp_compare(
  x,
  var_vec,
  scale = FALSE,
  sqrt = TRUE,
  trim = 0.1,
  trellis = TRUE,
  which_class = 2L,
  prob = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class randomForest.}

\item{var_vec}{Optional vector argument for reducing the number of
variables to consider and compare. Elements should be characters that
match column names from the data used to generate the model x.}

\item{scale}{For permutation based measures such as MSE or Accuracy, should
the measures be divided by their "standard errors"? Default is FALSE.}

\item{sqrt}{Boolean value indicating whether importance metrics should be
adjusted via a square root transformation. Default is True.}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from
each end of an individual PDP dataset before the trim-range is computed.
The default of 0.1 will be used when values of trim outside that range
are given.}

\item{trellis}{Logical indicating whether or not to generate trellis plots
as output for comparing PDPs. Default is TRUE.}

\item{which_class}{Integer specifying which column of the matrix of
predicted probabilities to use as the "focus" class. Default is to use
the first class. Only used for classification problems.}

\item{prob}{Logical indicating whether or not partial dependence for
classification problems should be returned on the probability scale,
rather than the centered logit. If FALSE, the partial dependence
function is on a scale similar to the logit. Default is TRUE.}

\item{...}{Other parameters to pass to the partial function.}
}
\value{
A list of partial dependence plots with adjusted y-axes so all
  are on an identical scale. This list includes a comparative facet plot and
  pdp importance values for assessing true affect of predictors on response.

A list of partial dependence plots with adjusted y-axes so all
  are on an identical scale. This list includes a comparative facet plot and
  pdp importance values for assessing true affect of predictors on response.
}
\description{
This function takes a randomForest object, generates partial
  dependence plots for predictors and converts them to small multiples for
  appropriate comparison. Output is a list containing a comparative grid
  of PDPs, individual partial dependence plots, and PDP-derived
  importance values for assessing effect of predictors on response.

This function takes a randomForest object, generates partial
  dependence plots for predictors and converts them to small multiples for
  appropriate comparison. Output is a list containing a comparative grid
  of PDPs, individual partial dependence plots, and PDP-derived
  importance values for assessing effect of predictors on response.
}
\examples{
mtcars.rf <- randomForest::randomForest(formula = mpg ~ ., data = mtcars)
car_pd <- pdp_compare(x = mtcars.rf)
car_pd$full
car_pd$imp
gridExtra::grid.arrange(car_pd$wt, car_pd$disp,
  car_pd$hp, car_pd$cyl, nrow = 2)
mtcars.rf <- randomForest::randomForest(formula = mpg ~ ., data = mtcars)
car_pd <- pdp_compare(x = mtcars.rf)
car_pd$full
car_pd$imp
gridExtra::grid.arrange(car_pd$wt, car_pd$disp,
  car_pd$hp, car_pd$cyl, nrow = 2)
}
